use shipped bootPatchLevel/vendorPatchLevel format
The documentation states that the format is YYYYMMDD for these fields in
order to distinguish between the split patch level. However, the actual
implementation on the Pixel 3 uses the YYYYMM format. For now, switch to
using the format that was shipped rather than what was documented. It
may become necessary to handle both formats if the documentation matches
how it was intended to be implementation.