recover from force stop cancelling jobs
If the user force stops the app or disables it, registered jobs are
cleared as part of the app entering the STOPPED state. If scheduled
remote verification is enabled, the job needs to be restored when the
user starts the app again. It isn't possible for the app to run before
the user has explicitly started it again and it wouldn't be desirable.