use row stride instead of image width for zxing
CameraX passes the raw data provided by the hardware which may be padded
for alignment. This depends on the combination of the chosen resolution
and the hardware. For example, a width of 960 pixels is evenly divisible
by 16/32/64 but 1080 pixels isn't evenly divisible and will often end up
having padding at the end of each row in the raw data.