add back timestamp verification for our own cert
This doesn't have much relevance to security due to verification of the
random challenge but we still want to verify it. This means attestation
will start failing due to a severely out-of-sync clock. That will often
break HTTPS so it tends to very noticeable and will often break remote
attestation even without this check since it can't make a connection.