remove Auditee self-check
For convenience, we implemented support for a basic verification of the
attestation on the Auditee side. However, this needs to go away as part
of giving the Auditor side including our remote attestation service more
visibility into why failures are happening.

Removing this has been planned for a long time but didn't happen until
now due to since it wasn't as trivial to remove before
228d62924005274220c23a8b53c222613b33c5f6.