add check to silence false positive lint warning
The BootReceiver class is explicitly not exported so this won't improve
anything but the lint expects it to be here. It didn't used to report an
issue so it was likely a regression in the check. It makes more sense to
add the check than silencing the lint in this case.