add new DEFLATE dictionary
This adds a new DEFLATE dictionary for protocol version 7.

The sample is a chain of 5 certificates from GrapheneOS 2025110600 on a
Pixel 9. The root is omitted and instead each of the valid roots is
included in the dictionary.

We no longer include both a TEE and StrongBox sample since TEE support
has been nearly entirely phased out and the savings from having a sample
with it in the dictionary are insignificant.

The main purpose of the new dictionary is to add the upcoming Google
root for when that starts being actively used in February 2026.